telegram-bot/requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# HTTP Requests
aiohttp>=3.8.0
requests>=2.31.0

# Database
supabase>=2.0.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Logging
structlog>=23.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0
uvloop==0.19.0 